+++ cli/js/compiler/ts_global.d.ts
