--- cli/js/mock_builtin.js
@@ -1,2 +1,0 @@
-// Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-export default undefined;
