+++ cli/js/symbols.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 import { internalSymbol } from "./internals.ts";
-import { customInspect } from "./console.ts";
+import { customInspect } from "./web/console.ts";
 
 /** Special Deno related symbols. */
 export const symbols = {
