+++ cli/js/web/performance.ts
@@ -1,5 +1,5 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import { now as opNow } from "./ops/timers.ts";
+import { now as opNow } from "../ops/timers.ts";
 
 export class Performance {
   /** Returns a current time from Deno's start in milliseconds.
