+++ cli/js/web/workers.ts
@@ -5,15 +5,15 @@ import {
   hostTerminateWorker,
   hostPostMessage,
   hostGetMessage
-} from "./ops/worker_host.ts";
-import { log } from "./util.ts";
-import { TextDecoder, TextEncoder } from "./web/text_encoding.ts";
+} from "../ops/worker_host.ts";
+import { log } from "../util.ts";
+import { TextDecoder, TextEncoder } from "./text_encoding.ts";
 /*
 import { blobURLMap } from "./web/url.ts";
 import { blobBytesWeakMap } from "./web/blob.ts";
 */
-import { Event } from "./web/event.ts";
-import { EventTarget } from "./web/event_target.ts";
+import { Event } from "./event.ts";
+import { EventTarget } from "./event_target.ts";
 
 const encoder = new TextEncoder();
 const decoder = new TextDecoder();
