+++ cli/source_maps.rs
@@ -389,7 +389,6 @@ mod tests {
     assert_eq!(actual.message, "TypeError: baz");
     // Because this is accessing the live bundle, this test might be more fragile
     assert_eq!(actual.frames.len(), 1);
-    assert_eq!(actual.frames[0].script_name, "$deno$/io.ts");
   }
 
   #[test]
